# server/src/main/resources/META-INF/openapi.yaml
...
openapi: 3.1.0
paths:
...
  /knowledge-units:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeUnit"
      summary: Get Knowledge Units
      tags:
      - Knowledge Unit Resource
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeUnitRequest"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Create Knowledge Unit
      tags:
      - Knowledge Unit Resource
  /knowledge-units/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeUnitRequest"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Update Knowledge Unit
      tags:
      - Knowledge Unit Resource
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeUnitWithDocumentsResponse"
      summary: Get Knowledge Unit With Documents
      tags:
      - Knowledge Unit Resource
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: No Content
        "200":
          description: OK
      summary: Delete Knowledge Unit
      tags:
      - Knowledge Unit Resource
  /knowledge-units/{id}/download:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
      summary: Download Knowledge Unit
      tags:
      - Knowledge Unit Manager Resource
components:
  schemas:
...
    Document:
      type: object
      required:
      - title
      - content
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        title:
          type: string
        content:
          type: string
        embedding: {}
        metadata:
          $ref: "#/components/schemas/JSONB"
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
        updatedAt:
          $ref: "#/components/schemas/LocalDateTime"
    DocumentRequest:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
    JSONB:
      type: object
      properties:
        data:
          type: string
    KnowledgeUnit:
      type: object
      required:
      - name
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
          maxLength: 255
        description:
          type: string
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
        updatedAt:
          $ref: "#/components/schemas/LocalDateTime"
    KnowledgeUnitRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        documents:
          type: array
          items:
            $ref: "#/components/schemas/DocumentRequest"
        newDocuments:
          type: array
          items:
            $ref: "#/components/schemas/DocumentRequest"
        addedDocumentsIds:
          type: array
          items:
            type: string
            format: uuid
            pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
        removedDocumentsIds:
          type: array
          items:
            type: string
            format: uuid
            pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
    LocalDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
    KnowledgeUnitWithDocumentsResponse:
      type: object
      properties:
        knowledgeUnit:
          $ref: "#/components/schemas/KnowledgeUnit"
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
info:
  title: Knowledge Manager API
  version: 1.0.0
  description: "The Knowledge Manager API provides endpoints for managing knowledge\
    \ units. A knowledge unit aggregates multi-dimensional information needed for\
    \ specific tasks, enabling cross-domain information reuse and contextual knowledge\
    \ assembly."
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value